@@ -10,12 +10,12 @@
 	},
 	"version": "0.1.3",
 	"scripts": {
-		"build": "rollup -c rollup.config.ts",
-		"lint": "eslint src --ext js,jsx,ts,tsx --report-unused-disable-directives --max-warnings 0"
+		"build": "rollup -c rollup.config.ts --configPlugin typescript",
+		"lint": "eslint --fix"
 	},
 	"main": "lib/index.cjs.js",
 	"module": "lib/index.esm.js",
-	"types": "lib/types",
+	"types": "lib/index.d.ts",
 	"browserslist": {
 		"production": [
 			">0.2%",
@@ -29,37 +29,34 @@
 		]
 	},
 	"devDependencies": {
-		"@babel/preset-env": "^7.22.9",
-		"@babel/preset-react": "^7.22.5",
-		"@rollup/plugin-commonjs": "^22.0.1",
+		"@babel/preset-env": "^7.25.8",
+		"@babel/preset-react": "^7.25.7",
+		"@eslint/compat": "^1.2.0",
+		"@eslint/eslintrc": "^3.1.0",
+		"@eslint/js": "^9.12.0",
+		"@rollup/plugin-commonjs": "^28.0.0",
 		"@rollup/plugin-image": "^3.0.3",
-		"@rollup/plugin-node-resolve": "^13.3.0",
-		"@rollup/plugin-terser": "^0.4.3",
-		"@types/node": "^20.11.30",
-		"@types/react": "^18.0.15",
-		"@types/react-dom": "^18.0.6",
-		"@typescript-eslint/eslint-plugin": "^5.62.0",
-		"@vitejs/plugin-react": "^4.0.3",
+		"@rollup/plugin-node-resolve": "^15.3.0",
+		"@rollup/plugin-terser": "^0.4.4",
+		"@rollup/plugin-typescript": "^12.1.0",
+		"@types/node": "^22.7.5",
+		"@types/react": "^18.3.11",
+		"@types/react-dom": "^18.3.1",
+		"@typescript-eslint/eslint-plugin": "^8.9.0",
+		"@typescript-eslint/parser": "^8.9.0",
+		"@vitejs/plugin-react": "^4.3.2",
 		"browserslist-to-esbuild": "^2.1.1",
-		"eslint": "^8.45.0",
-		"eslint-config-react-app": "^7.0.1",
-		"eslint-config-standard-with-typescript": "^36.1.0",
-		"eslint-plugin-import": "^2.27.5",
-		"eslint-plugin-n": "^16.0.1",
-		"eslint-plugin-promise": "^6.1.1",
-		"eslint-plugin-react": "^7.30.1",
-		"eslint-plugin-react-hooks": "^4.6.0",
-		"eslint-plugin-react-refresh": "^0.4.3",
-		"prettier": "^2.7.1",
-		"react": "^18.2.0",
-		"react-dom": "^18.2.0",
-		"rollup": "^2.75.7",
-		"rollup-plugin-typescript2": "^0.32.1",
-		"typescript": "^4.7.4",
-		"vite": "^4.4.4"
+		"eslint": "^9.12.0",
+		"eslint-plugin-react": "^7.37.1",
+		"globals": "^15.11.0",
+		"prettier": "^3.3.3",
+		"rollup": "^4.24.0",
+		"tslib": "^2.8.0",
+		"typescript": "^5.6.3",
+		"vite": "^5.4.9"
 	},
 	"peerDependencies": {
-		"react": "^17.0.2 || ^18",
-		"react-dom": "^17.0.2 || ^18"
+		"react": "^18.3.1",
+		"react-dom": "^18.3.1"
 	}
 }
