@@ -1,12 +1,15 @@
 import commonjs from '@rollup/plugin-commonjs'
 import image from '@rollup/plugin-image'
-import pkg from './package.json'
+import pkg from './package.json' assert { type: 'json' }
 import resolve from '@rollup/plugin-node-resolve'
 import terser from '@rollup/plugin-terser'
-import typescript from 'rollup-plugin-typescript2'
+import typescript from '@rollup/plugin-typescript'
 
 const config = {
-	external: ['react', 'react-dom'],
+	external: [
+		...Object.keys(pkg.devDependencies),
+		...Object.keys(pkg.peerDependencies)
+	],
 	input: 'src/index.ts',
 	output: [
 		{
@@ -22,9 +25,7 @@ const config = {
 		resolve({ browser: true }),
 		commonjs(),
 		image(),
-		typescript({
-			useTsconfigDeclarationDir: true
-		}),
+		typescript(),
 		terser()
 	]
 }
